pyproject.toml
[project]
name = "be-server-for-fe-final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.120.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.3",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
